pyproject.toml
[tool.poetry]
name = "resize-img-script"
version = "0.1.0"
description = ""
authors = ["Aleksandra <abaibik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
tqdm = "^4.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
